requirements.txt
boto3==1.26.95
botocore==1.29.95
click==8.1.3
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
py4j==0.10.9.5
pyspark==3.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
